Added requirements.txt and updated main.py
# Added yaspin support
Now, there is a fully functioning progress indicator that uses the yaspin module.

# Removed date modified check
Previously, the script would check if the local data stored was already today's date. However, if the data was updated through the day and the source file was updated, the script would not download the newest version.

I have removed this for simplicity purposes.

# Added requirements.txt
I'm not sure if its the conventional method of writing a requirements file, but it serves to show what modules the user may need to install.